pyproject.toml
[project]
name = "MasterTutor"
version = "0.1.0"
description = "Multi-agent application template using LangGraph and LangChain"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=1.0.3",
    "langchain>=1.0.0",
    "langchain-core>=1.0.4",
    "langchain-openai>=1.0.0",
    "langchain-anthropic>=1.0.0",
    "langchain-google-genai>=3.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "langsmith>=0.1.0",
    "tenacity>=8.2.0",
    "tiktoken>=0.7.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
extend = "ruff.toml"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
